+++ src/extension.ts
@@ -168,24 +168,16 @@ function openAttachment(option: any): void {
 function initZoteroDb(): ZoteroDatabase {
     const config = vscode.workspace.getConfiguration('zotero');
     let zoteroDbPath = config.get<string>('zoteroDbPath', '~/Zotero/zotero.sqlite');
-    let betterBibtexDbPath = config.get<string>('betterBibtexDbPath', '~/Zotero/better-bibtex.sqlite');
     zoteroDbPath = expandPath(zoteroDbPath);
-    betterBibtexDbPath = expandPath(betterBibtexDbPath);
 
     const zoteroDbExists = existsSync(zoteroDbPath);
-    const betterBibtexDbExists = existsSync(betterBibtexDbPath);
 
     if (!zoteroDbExists) {
         vscode.window.showErrorMessage(`Zotero database not found at path: ${zoteroDbPath}`);
         throw new Error(`Zotero database not found at path: ${zoteroDbPath}`);
     }
 
-    return new ZoteroDatabase({
-        zoteroDbPath: zoteroDbPath,
-        betterBibtexDbPath: betterBibtexDbPath,
-        // if better-bibtex.sqlite exists, we will use it to get citation keys (legacy method)
-        betterBibtexLegacy: betterBibtexDbExists 
-    });
+    return new ZoteroDatabase(zoteroDbPath);
 }